- 소스정리 - 기구물에 실장 후 캘리브레이션 - 추 후 가속도 값을 변경하여 다시 캘리브레이션 해야함

+++ Inc/mpu6050_dmp.h
@@ -373,7 +373,7 @@ extern float Pitch,Roll,Roll_reverse,Yaw, Rangle, Pangle;
 extern float base_pitch,base_roll,base_yaw,base_roll_reverse;
 extern float dqw, dqx, dqy, dqz;
 extern uint8_t Cal_done;
-extern uint8_t ledPos;
+extern float ledPos;
 extern float targetLedPos;
 extern float targetAnglel;
